feat(layout grid): use Context in grid so React.Fragments can be used (#3264)
Change grid to use context so that non-cell elements can be used as children to the grid. This will allow you to split up cells into subcomponents that have a react fragment wrapped around them, or be wrapped in HoCs/other wrapped components that still render Cells.

This change should be entirely backwards compatible with any existing code that uses Baseweb.